Fix boostlib check in configure.ac
The macro usage here was just totally incorrect - not only did the
check not work (i.e. the configure script passed even if boostlib was
not found), but when boostlib was found, it caused a file named = to
be created.  Both these issues have been resolved by correctly quoting
the arguments to the boost check macros, as well as adding an additional
blank argument to ensure the AC_MSG_ERROR is evaluated in the "not found"
case rather than the "found" case.